refactor(cove-tokio): move timeout ext and root tokio init
